libgomp: Device load_image - minor num-funcs/vars check improvement
When debugging a linker issue, leading to a mismatch in the number of
host/device functions, I was surprised by seeing one additional entry.
Well, it turned out to be due to the ICV variable.

This patch makes it more consistent. The "+1" is returned since
r12-2769-g0bac793ed6bad2 (for the on-device omp_get_device_num),
extended in r13-2545-g9f2fca56593a2b for a struct to support more ICV
variables on the devices [to handle OMP_..._DEV environment variables].

As the value is returned unconditionally, it makes sense to use it both
for the expected-value diagnostic and for the condition further below.

Comments, suggestions, remarks?

Tobias

PS: Alternative would be to make the plugin's value depend on whether
the data was loaded. But that would make the number-of-entries assert
weaker and might cause corner-case issues when a slightly older libgomp
plugin is used with the updated libgomp.so. Thus, I have settled for the
attached variant.

libgomp: Device load_image - improve minor num-funcs/vars check

The run time library loads the offload functions and variable and optionally
the ICV variable and returns the number of loaded items, which has to match
the host side. The plugin returns "+1" (since GCC 12) for the ICV variable
entry, independently whether it was loaded or not, but the var's value
(start == end == 0) can be used to detect when this failed.

Thus, we can tighten the assert check - which this commit does together with
making the output less surprising - and simplify the condition further below.

libgomp/ChangeLog:

	* plugin/plugin-gcn.c (GOMP_OFFLOAD_load_image): If ICV variable
	is is not available, decrement other_count and thus the return value.
	* plugin/plugin-nvptx.c (GOMP_OFFLOAD_load_image): Likewise.
	* target.c (gomp_load_image_to_device): Extend fatal-error message;
	simplify a condition.

 libgomp/target.c | 78 +++++++++++++++++++++++++-------------------------------
 1 file changed, 35 insertions(+), 43 deletions(-)