libcpp: Improve location for macro names [PR66290]
This backport for gcc-13 is the first of two required for the
g++.dg/pch/line-map-3.C test to stop hitting a variant of the known
problem mentioned in that testcase: on riscv64-elf and riscv32-elf,
after restoring the PCH, the location of the macros is mentioned as if
they were on line 3 rather than 2, so even the existing xfails fail.  I
think this might be too much to backport, and I'm ready to use an xfail
instead, but since this would bring more predictability, I thought I'd
ask whether you'd find this backport acceptable.

Regstrapped on x86_64-linux-gnu, along with other backports, and tested
manually on riscv64-elf.  Ok to install?

From: Lewis Hyatt <lhyatt@gmail.com>

When libcpp reports diagnostics whose locus is a macro name (such as for
-Wunused-macros), it uses the location in the cpp_macro object that was
stored by _cpp_new_macro. This is currently set to pfile->directive_line,
which contains the line number only and no column information. This patch
changes the stored location to the src_loc for the token defining the macro
name, which includes the location and range information.

libcpp/ChangeLog:

	PR c++/66290
	* macro.cc (_cpp_create_definition): Add location argument.
	* internal.h (_cpp_create_definition): Adjust prototype.
	* directives.cc (do_define): Pass new location argument to
	_cpp_create_definition.
	(do_undef): Stop passing inferior location to cpp_warning_with_line;
	the default from cpp_warning is better.
	(cpp_pop_definition): Pass new location argument to
	_cpp_create_definition.
	* pch.cc (cpp_read_state): Likewise.

gcc/testsuite/ChangeLog:

	PR c++/66290
	* c-c++-common/cpp/macro-ranges.c: New test.
	* c-c++-common/cpp/line-2.c: Adapt to check for column information
	on macro-related libcpp warnings.
	* c-c++-common/cpp/line-3.c: Likewise.
	* c-c++-common/cpp/macro-arg-count-1.c: Likewise.
	* c-c++-common/cpp/pr58844-1.c: Likewise.
	* c-c++-common/cpp/pr58844-2.c: Likewise.
	* c-c++-common/cpp/warning-zero-location.c: Likewise.
	* c-c++-common/pragma-diag-14.c: Likewise.
	* c-c++-common/pragma-diag-15.c: Likewise.
	* g++.dg/modules/macro-2_d.C: Likewise.
	* g++.dg/modules/macro-4_d.C: Likewise.
	* g++.dg/modules/macro-4_e.C: Likewise.
	* g++.dg/spellcheck-macro-ordering.C: Likewise.
	* gcc.dg/builtin-redefine.c: Likewise.
	* gcc.dg/cpp/Wunused.c: Likewise.
	* gcc.dg/cpp/redef2.c: Likewise.
	* gcc.dg/cpp/redef3.c: Likewise.
	* gcc.dg/cpp/redef4.c: Likewise.
	* gcc.dg/cpp/ucnid-11-utf8.c: Likewise.
	* gcc.dg/cpp/ucnid-11.c: Likewise.
	* gcc.dg/cpp/undef2.c: Likewise.
	* gcc.dg/cpp/warn-redefined-2.c: Likewise.
	* gcc.dg/cpp/warn-redefined.c: Likewise.
	* gcc.dg/cpp/warn-unused-macros-2.c: Likewise.
	* gcc.dg/cpp/warn-unused-macros.c: Likewise.

(cherry picked from commit 4f3be7cbebce8ec9e0c5d9340b2772581454b862)