rs6000, Update comment for the __builtin_vsx_vper*, built-ins.
GCC maintainers:

The patch expands an existing comment to document that the duplicates are covered by an overloaded built-in.  I am wondering if we should just go ahead and remove the duplicates?

The patch has been tested on Power 10 with no regressions.

Please let me know if this patch is acceptable for mainline.  Thanks.

                      Carl

-----------------------------------------------------
rs6000, Update comment for the __builtin_vsx_vper* built-ins.

There is a comment about the __builtin_vsx_vper* built-ins being
duplicates of the __builtin_altivec_* built-ins.  The note says we
should consider deprecation/removeal of the __builtin_vsx_vper*.  Add a
note that the _builtin_vsx_vper* built-ins are covered by the overloaded
vec_perm built-ins which use the __builtin_altivec_* built-in definitions.

gcc/ChangeLog:
	* config/rs6000/rs6000-builtins.def ( __builtin_vsx_vperm_*):
	Add comment to existing comment about the built-ins.