ipa: Create indirect call edges also for lambdas
Hi,

Even though we have had code to handle creation of indirect call graph
edges (so that these calls can than be made direct as part of IPA-CP
and inlining and eventually also inlined) for C++ member pointers for
many years, this code expects the member pointers to be structures
passed by value.  In PR 108802 it turned out that for lambdas these
are passed by reference.  This patch adjusts the code for that so that
small lambdas are also inlineable without depending on early inlining.

Bootstrapped and LTO bootstrapped on x86_64-linux.  This is technically
a regression against GCC 10.  OK for master even now?

Thanks,

Martin

gcc/ChangeLog:

2024-02-20  Martin Jambor  <mjambor@suse.cz>

	PR ipa/108802
	* ipa-prop.cc (ipa_get_stmt_member_ptr_load_param): Also recognize
	loads from a pointer parameter.
	(ipa_analyze_indirect_call_uses): Likewise.

gcc/testsuite/ChangeLog:

2024-02-20  Martin Jambor  <mjambor@suse.cz>

	PR ipa/108802
	* g++.dg/ipa/pr108802.C: New test.