c++/modules: Stream definitions for implicit instantiations [PR114170]
Bootstrapped and regtested on x86_64-pc-linux-gnu, OK for trunk?

-- >8 --

An implicit instantiation has an initializer depending on whether
DECL_INITIALIZED_P is set (like normal VAR_DECLs) which needs to be
written to ensure that consumers of header modules properly emit
definitions for these instantiations. This patch ensures that we
correctly fallback to checking this flag when DECL_INITIAL is not set
for a template instantiation.

As a drive-by fix, also ensures that the count of initializers matches
the actual number of initializers written. This doesn't seem to be
necessary for correctness in the current testsuite, but feels wrong and
makes debugging harder when initializers aren't properly written for
other reasons.

	PR c++/114170

gcc/cp/ChangeLog:

	* module.cc (has_definition): Fall back to DECL_INITIALIZED_P
	when DECL_INITIAL is not set on a template.
	(module_state::write_inits): Only increment count when
	initializers are actually written.

gcc/testsuite/ChangeLog:

	* g++.dg/modules/var-tpl-2_a.H: New test.
	* g++.dg/modules/var-tpl-2_b.C: New test.

Signed-off-by: Nathaniel Shead <nathanieloshead@gmail.com>